requirements.txt
# Core agent orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# HuggingFace (primary LLM + embeddings, replaces Gemini + Groq)
langchain-huggingface>=0.1.0
huggingface-hub>=0.24.0
sentence-transformers>=3.0.0

# Data sources
arxiv>=2.1.0
tavily-python>=0.3.0

# Vector store
chromadb>=0.5.0

# Structured outputs
pydantic>=2.0.0

# MCP server
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
